# contracts/openapi-v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CampusMarket API",
    "description": "API HTTP/JSON de CampusMarket para crear y consultar publicaciones. El contrato versionado es contracts/openapi-v2.json.",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Entorno local"
    }
  ],
  "paths": {
    "/publicaciones": {
      "get": {
        "tags": [
          "publicaciones"
        ],
        "summary": "Listar publicaciones",
        "operationId": "listarPublicaciones",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Publicacion"
                  },
                  "type": "array",
                  "title": "Response Listarpublicaciones"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "publicaciones"
        ],
        "summary": "Crear una publicación",
        "operationId": "crearPublicacion",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicacionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Publicacion"
                }
              }
            }
          },
          "503": {
            "description": "Persistencia temporalmente no disponible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/publicaciones/mias": {
      "get": {
        "tags": [
          "publicaciones"
        ],
        "summary": "Listar publicaciones de un propietario",
        "operationId": "listarMisPublicaciones",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PublicacionGestion"
                  },
                  "type": "array",
                  "title": "Response Listarmispublicaciones"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/publicaciones/{publication_id}": {
      "put": {
        "tags": [
          "publicaciones"
        ],
        "summary": "Editar una publicación propia",
        "operationId": "editarPublicacion",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "publication_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Publication Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublicacionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Publicacion"
                }
              }
            }
          },
          "404": {
            "description": "Publicación no encontrada para el propietario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Persistencia temporalmente no disponible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "publicaciones"
        ],
        "summary": "Eliminar una publicación propia",
        "operationId": "eliminarPublicacion",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "publication_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Publication Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Publicación no encontrada para el propietario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Persistencia temporalmente no disponible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/publicaciones/{publication_id}/estado": {
      "patch": {
        "tags": [
          "publicaciones"
        ],
        "summary": "Cambiar disponibilidad de una publicación propia",
        "operationId": "cambiarEstadoPublicacion",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "publication_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Publication Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstadoPublicacionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Publicacion"
                }
              }
            }
          },
          "404": {
            "description": "Publicación no encontrada para el propietario",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Persistencia temporalmente no disponible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/publicaciones/{publication_id}/imagenes": {
      "post": {
        "tags": [
          "publicaciones"
        ],
        "summary": "Subir una imagen de una publicación",
        "operationId": "subirImagenPublicacion",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "publication_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Publication Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_subirImagenPublicacion"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicacionImagen"
                }
              }
            }
          },
          "400": {
            "description": "Imagen inválida o máximo alcanzado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Publicación no encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Persistencia temporalmente no disponible",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_subirImagenPublicacion": {
        "properties": {
          "archivo": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Archivo"
          }
        },
        "type": "object",
        "required": [
          "archivo"
        ],
        "title": "Body_subirImagenPublicacion"
      },
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorResponse"
      },
      "EstadoPublicacionUpdate": {
        "properties": {
          "estado_publicacion": {
            "type": "string",
            "enum": [
              "disponible",
              "reservado",
              "vendido"
            ],
            "title": "Estado Publicacion"
          }
        },
        "type": "object",
        "required": [
          "estado_publicacion"
        ],
        "title": "EstadoPublicacionUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Publicacion": {
        "properties": {
          "titulo": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Titulo"
          },
          "descripcion": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Descripcion"
          },
          "precio": {
            "type": "number",
            "maximum": 9999999999.99,
            "exclusiveMinimum": 0.0,
            "title": "Precio"
          },
          "modalidad": {
            "type": "string",
            "enum": [
              "venta",
              "alquiler"
            ],
            "title": "Modalidad"
          },
          "estado": {
            "type": "string",
            "enum": [
              "nuevo",
              "usado",
              "reacondicionado"
            ],
            "title": "Estado"
          },
          "estado_publicacion": {
            "type": "string",
            "enum": [
              "disponible",
              "reservado",
              "vendido"
            ],
            "title": "Estado Publicacion",
            "default": "disponible"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "propietario_id": {
            "type": "integer",
            "title": "Propietario Id"
          },
          "visible": {
            "type": "boolean",
            "title": "Visible",
            "default": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "titulo",
          "descripcion",
          "precio",
          "modalidad",
          "estado",
          "id",
          "propietario_id"
        ],
        "title": "Publicacion"
      },
...
      "PublicacionCreate": {
        "properties": {
          "titulo": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Titulo"
          },
          "descripcion": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Descripcion"
          },
          "precio": {
            "type": "number",
            "maximum": 9999999999.99,
            "exclusiveMinimum": 0.0,
            "title": "Precio"
          },
          "modalidad": {
            "type": "string",
            "enum": [
              "venta",
              "alquiler"
            ],
            "title": "Modalidad"
          },
          "estado": {
            "type": "string",
            "enum": [
              "nuevo",
              "usado",
              "reacondicionado"
            ],
            "title": "Estado"
          },
          "estado_publicacion": {
            "type": "string",
            "enum": [
              "disponible",
              "reservado",
              "vendido"
            ],
            "title": "Estado Publicacion",
            "default": "disponible"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "titulo",
          "descripcion",
          "precio",
          "modalidad",
          "estado"
        ],
        "title": "PublicacionCreate"
      },
      "PublicacionGestion": {
        "properties": {
          "titulo": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Titulo"
          },
          "descripcion": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Descripcion"
          },
          "precio": {
            "type": "number",
            "maximum": 9999999999.99,
            "exclusiveMinimum": 0.0,
            "title": "Precio"
          },
          "modalidad": {
            "type": "string",
            "enum": [
              "venta",
              "alquiler"
            ],
            "title": "Modalidad"
          },
          "estado": {
            "type": "string",
            "enum": [
              "nuevo",
              "usado",
              "reacondicionado"
            ],
            "title": "Estado"
          },
          "estado_publicacion": {
            "type": "string",
            "enum": [
              "disponible",
              "reservado",
              "vendido"
            ],
            "title": "Estado Publicacion",
            "default": "disponible"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "propietario_id": {
            "type": "integer",
            "title": "Propietario Id"
          },
          "visible": {
            "type": "boolean",
            "title": "Visible",
            "default": true
          },
          "imagenes": {
            "items": {
              "$ref": "#/components/schemas/PublicacionImagen"
            },
            "type": "array",
            "title": "Imagenes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "titulo",
          "descripcion",
          "precio",
          "modalidad",
          "estado",
          "id",
          "propietario_id"
        ],
        "title": "PublicacionGestion"
      },
      "PublicacionImagen": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "publicacion_id": {
            "type": "integer",
            "title": "Publicacion Id"
          },
          "imagen_url": {
            "type": "string",
            "title": "Imagen Url"
          },
          "orden": {
            "type": "integer",
            "title": "Orden"
          },
          "es_principal": {
            "type": "boolean",
            "title": "Es Principal"
          }
        },
        "type": "object",
        "required": [
          "id",
          "publicacion_id",
          "imagen_url",
          "orden",
          "es_principal"
        ],
        "title": "PublicacionImagen"
      },
      "PublicacionUpdate": {
        "properties": {
          "titulo": {
            "type": "string",
            "maxLength": 100,
            "minLength": 3,
            "title": "Titulo"
          },
          "descripcion": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Descripcion"
          },
          "precio": {
            "type": "number",
            "maximum": 9999999999.99,
            "exclusiveMinimum": 0.0,
            "title": "Precio"
          },
          "modalidad": {
            "type": "string",
            "enum": [
              "venta",
              "alquiler"
            ],
            "title": "Modalidad"
          },
          "estado": {
            "type": "string",
            "enum": [
              "nuevo",
              "usado",
              "reacondicionado"
            ],
            "title": "Estado"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "titulo",
          "descripcion",
          "precio",
          "modalidad",
          "estado"
        ],
        "title": "PublicacionUpdate"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}